creating questions.js
